changed names to sti_resolv
